\name{engel}
\alias{engel}
\title{Engel Data}
\description{
Engel Food Expenditure Data used in Koenker and Bassett(1982).
This is a regression data set consisting of 235 observations on 
income and expenditure on food for Belgian working class households.
}
\usage{data(engel)}
\format{A data frame containing 235 observations on 2 variables
\describe{
\item{x}{ "Annual Household Income in Belgian Francs"}
\item{y}{"Annual Household Food Expenditure in Belgian Francs"}
}
}
\references{
Koenker, R. and Bassett, G (1982) Robust Tests of Heteroscedasticity based on
Regression Quantiles, Econometrica, 50,43--61.}
\keyword{datasets}
