\name{FAQ}
\alias{FAQ}
\alias{ChangeLog}
\title{FAQ and ChangeLog  of a package}
\description{ Show the FAQ or ChangeLog of a specified package  }
\usage{
FAQ(pkg = "quantreg")
ChangeLog(pkg = "quantreg")
}
\arguments{
  \item{pkg}{ Package Name }
}
\details{
Assumes that the FAQ and/or ChangeLog files exist in the proper "inst" directory.
}
\value{
Has only the side effect of showing the files on the screen.  
}
\keyword{ documentation }
