\name{latex}
\alias{latex}
\title{ ~~Generic method to make latex fragments  ~~ }
\description{
	This method has only one instance (for the moment) and that is
	to make the output of table.rq() into a latex table.
}
\usage{
latex(x, ...)
}
\arguments{
  \item{x}{ ~~Object \code{x} to be converted to latex~~ }
  \item{\dots}{ ~~Other arguments \code{\dots} here~~ }
}
\details{
	see the example function latex.table.rq
}
\value{
	writes a file containing the latex code
}
\references{ ~put references to the literature/web site here ~ }
\author{ Roger Koenker }


\seealso{  \code{\link{latex.table.rq}}, ~~~ }

\keyword{ utilities }% at least one, from doc/KEYWORDS
