\name{MelTemp}
\alias{MelTemp}
\docType{data}
\title{Daily maximum temperatures in Melbourne, Australia}
\description{
     Daily maximum temperatures in Melbourne, Australia, from
     1981-1990. Leap days have been omitted.
}
\usage{data(MelTemp)}
\format{Time series of frequency 365}

\source{
Hyndman, R.J., Bashtannyk, D.M. and Grunwald, G.K. (1996)
     "Estimating and visualizing conditional densities". _Journal of
     Computational and Graphical Statistics_, *5*, 315-336.
}
\examples{
data(MelTemp)
demo(Mel)
}
\keyword{datasets}
