\name{Patacamaya}
\non_function{}
\title{Patacamaya: Greenhouse experiment at Patacamaya  }
\description{
  This dataset contains the following variables: \cr \cr
   y       Daily (nightly) minimum temperature \cr
           inside greenhouse, degrees centigrades. \cr
   a       Minimum ambiental temperature. \cr
   inv     factor for 12 greenhouses.  \cr
   tipo    factor for 4 different greenhouse \cr 
           constructions. \cr \cr

  Data due to Cecilia Martinez, from an experiment at Patacamaya, the Bolivian altiplano.  
}
