
\name{plot.KhmaladzeTest}
\alias{plot.KhmaladzeTest}
\title{ Plot a KhmaladzeTest object}
\description{
Plot an object generated by KhmaladzeTest
}
\usage{
\method{plot}{KhmaladzeTest}(x, ...)
}
\arguments{
\item{x}{
  Object returned from KhmaladzeTest representing the fit of the model.
 }
\item{...}{
  Optional arguments. 
 }
}
\seealso{  \code{\link{KhmaladzeTest}}}

\keyword{ regression}
