\name{plot.rq.process}
\alias{plot.rq.process}
\title{ plot the coordinates of the quantile regression process}
\description{ Function to plot quantile regression  process.  }
\usage{ plot.rq.process(x, nrow=3, ncol=2, ...) }
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ an object produced by rq() fitting  }
  \item{nrow}{ rows in mfrow }
  \item{ncol}{ columns in mfrow}
  \item{...}{ optional arguments to plot}
}
\author{ Roger Koenker rkoenker@uiuc.edu}
\seealso{ \code{\link{rq}}}

\keyword{ regression }
