\name{print.KhmaladzeTest}
\alias{print.KhmaladzeTest}
\title{ Print a KhmaladzeTest object}
\description{
Print an object generated by KhmaladzeTest
}
\usage{
\method{print}{KhmaladzeTest}(x, ...)
}
\arguments{
\item{x}{
  Object returned from KhmaladzeTest representing the fit of the model.
 }
\item{...}{
  Optional arguments. 
 }
}
\seealso{  \code{\link{KhmaladzeTest}}}

\keyword{ regression}
