\name{print.rq}
\alias{print.rq}
\title{ Print an rq object}
\description{
Print an object generated by rq
}
\usage{
print.rq(x, ...)
}
\arguments{
\item{x}{
  Object returned from rq representing the fit of the model.
 }
\item{...}{
  Optional arguments passed to fitting routines
 }
}
\seealso{  \code{\link{rq}}}

\keyword{ regression}
