\name{latex}
\alias{latex}
\title{ Make a latex version of an R object }
\description{
Generic function for converting an  \R object into a latex file.
}
\usage{
latex(x, ...)
}
\arguments{
  \item{x}{ \code{x} is an \R object }
  \item{\dots}{ \code{\dots} optional arguments }
}
\seealso{  \code{\link{latex.table}}, \code{\link{latex.summary.rqs}}
}
\keyword{IO}

