\name{print.rq}
\alias{print.rq}
\alias{print.rqs}
\title{ Print an rq object}
\description{
Print an object generated by rq
}
\usage{
\method{print}{rq}(x, ...)
\method{print}{rqs}(x, ...)
}
\arguments{
\item{x}{
  Object returned from rq representing the fit of the model.
 }
\item{...}{
  Optional arguments. 
 }
}
\seealso{  \code{\link{rq}}}

\keyword{ regression}
