\name{table.rq}
\alias{table.rq}
\alias{latex.table.rq}
\alias{plot.table.rq}
\title{
  Table of Quantile Regression Results
}
\description{
  Defunct Function to produce a table of quantile regression results for a group
  of specified quantiles.  See \code{rq} which now permits multiple taus.
}
\usage{
table.rq(x, \dots)
}
\arguments{
  \item{x}{
	input
  }
  \item{...}{
    other optional arguments 
  }
}
\value{
	None.
}
\seealso{ \code{\link{rq}}, }
\keyword{regression}
